+++ back-bone-service/src/main/java/com/local/carcarecenter/controller/ItemController.java
@@ -24,7 +24,7 @@ public class ItemController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ItemViewModel> getById(@PathVariable Integer id) throws EntityNotFoundExecution {
+    public ResponseEntity<ItemViewModel> getById(@PathVariable Long id) throws EntityNotFoundExecution {
         var result = service.getById(id);
         return ResponseEntity.ok().body(result);
     }
@@ -35,13 +35,13 @@ public class ItemController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ItemViewModel> update(@PathVariable Integer id, @Valid @RequestBody ItemInputModel entity) throws EntityNotFoundExecution {
+    public ResponseEntity<ItemViewModel> update(@PathVariable Long id, @Valid @RequestBody ItemInputModel entity) throws EntityNotFoundExecution {
         var result = service.update(id, entity);
         return ResponseEntity.ok().body(result);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> delete(@PathVariable Integer id) throws EntityNotFoundExecution {
+    public ResponseEntity<Boolean> delete(@PathVariable Long id) throws EntityNotFoundExecution {
         service.getById(id);
         service.delete(id);
