+++ back-bone-service/src/main/java/com/local/carcarecenter/repository/RepairRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface RepairRepository extends JpaRepository<Repair, Integer> {
+public interface RepairRepository extends JpaRepository<Repair, Long> {
 }
