+++ back-bone-service/src/main/java/com/local/carcarecenter/service/AbstractBaseService.java
@@ -9,9 +9,9 @@ import java.util.List;
  * @param <V> ViewModel
  */
 public interface AbstractBaseService<I, V> {
-   V getById(Integer id) throws EntityNotFoundExecution;
+   V getById(Long id) throws EntityNotFoundExecution;
    List<V> getAll();
    V save(I model) throws EntityNotFoundExecution;
-   V update(Integer id, I model) throws EntityNotFoundExecution;
-   boolean delete(Integer id) throws EntityNotFoundExecution;
+   V update(Long id, I model) throws EntityNotFoundExecution;
+   boolean delete(Long id) throws EntityNotFoundExecution;
 }
